Add Turbine::Collection.
Collection is the start of a DSL for traversing through nodes and edges
and allows you to chain expressions together.

For example, from a node, we can select all nodes connected by an edge
with the :gas label:

  node.out(:gas)

... and we can select *their* electricity outputs:

  node.out(:gas).out(:electricity)

... or we can select their electricity outputs, but only as long as they
don't also output hot water... and then select *their* output nodes:

  without_hot_water = node.out(:gas).out(:electricity).select do |node|
    node.out_edges(:hot_water).empty?
  end

  without_hot_water.out

... and so on, and so on.

Like jQuery collection, it doesn't matter if any of the expressions
match nothing, one node, or a thousand nodes.